refactor: auth 관련 에러 처리 auth 도메인으로 이동 (#8)
