+++ src/main/java/com/chungbazi/server/domain/auth/exception/code/AuthErrorCode.java
@@ -1,4 +1,4 @@
-package com.chungbazi.server.global.common.code.status;
+package com.chungbazi.server.domain.auth.exception.code;
 
 import com.chungbazi.server.global.common.code.BaseErrorCode;
 import com.chungbazi.server.global.common.code.ErrorReasonDto;
@@ -8,12 +8,12 @@ import org.springframework.http.HttpStatus;
 
 @Getter
 @AllArgsConstructor
-public enum ErrorStatus implements BaseErrorCode {
-    //일반적인 응답
-    _INTERNAL_SERVER_ERROR(HttpStatus.INTERNAL_SERVER_ERROR, "COMMON500", "서버 에러, 관리자에게 문의 바랍니다."),
-    _BAD_REQUEST(HttpStatus.BAD_REQUEST, "COMMON400", "잘못된 요청입니다."),
-    _UNAUTHORIZED(HttpStatus.UNAUTHORIZED, "COMMON401", "인증이 필요합니다."),
-    _FORBIDDEN(HttpStatus.FORBIDDEN, "COMMON403", "금지된 요청입니다."),
+public enum AuthErrorCode implements BaseErrorCode {
+
+    INVALID_TOKEN(HttpStatus.UNAUTHORIZED, "AUTH401_1", "유효하지 않은 토큰입니다."),
+    EXPIRED_TOKEN(HttpStatus.UNAUTHORIZED, "AUTH401_2", "만료된 토큰입니다."),
+    KAKAO_API_ERROR(HttpStatus.BAD_GATEWAY, "AUTH502_1", "카카오 API 요청에 실패했습니다."),
+    KAKAO_REQUIRED_INFO_MISSING(HttpStatus.BAD_REQUEST, "AUTH400_1", "카카오 사용자 필수 정보가 누락되었습니다."),
     ;
 
     private final HttpStatus httpStatus;
